add BeanFather and refactor Beans
