+++ src/main/java/com/epam/spring/homework2/beans/BeanA.java
@@ -3,35 +3,9 @@ package com.epam.spring.homework2.beans;
 import org.springframework.stereotype.Component;
 
 @Component
-public class BeanA {
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-    private String name;
-    private String value;
+public class BeanA extends BeanFather {
 
     public BeanA(String name, String value) {
-        this.name = name;
-        this.value = value;
-    }
-
-    public String getValue() {
-        return value;
-    }
-
-    public void setValue(String value) {
-        this.value = value;
-    }
-
-    @Override
-    public String toString() {
-        return "BeanA{" +
-                "name='" + name + '\'' +
-                ", value='" + value + '\'' +
-                '}';
+        super(name, value);
     }
 }
