+++ src/main/java/com/epam/spring/homework2/beans/BeanE.java
@@ -3,35 +3,9 @@ package com.epam.spring.homework2.beans;
 import org.springframework.stereotype.Component;
 
 @Component
-public class BeanE {
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-    private String name;
-    private String value;
+public class BeanE extends BeanFather {
 
     public BeanE(String name, String value) {
-        this.name = name;
-        this.value = value;
-    }
-
-    public String getValue() {
-        return value;
-    }
-
-    public void setValue(String value) {
-        this.value = value;
-    }
-
-    @Override
-    public String toString() {
-        return "BeanE{" +
-                "name='" + name + '\'' +
-                ", value='" + value + '\'' +
-                '}';
+        super(name, value);
     }
 }
